moved parseTime function to fluxutil
