+++ Density/app/src/main/java/org/cornelldti/density/density/network/JsonParser.kt
@@ -3,6 +3,7 @@ package org.cornelldti.density.density.network
 import android.text.format.DateFormat
 import org.cornelldti.density.density.DensityApplication
 import org.cornelldti.density.density.data.*
+import org.cornelldti.density.density.util.FluxUtil
 import org.json.JSONArray
 import org.json.JSONException
 import java.text.SimpleDateFormat
@@ -94,7 +95,7 @@ object JsonParser {
                 val segment = hours.getJSONObject(i).getJSONObject("dailyHours")
                 val start = segment.getLong("startTimestamp")
                 val end = segment.getLong("endTimestamp")
-                operatingHours.add(parseTime(start) + " – " + parseTime(end))
+                operatingHours.add(FluxUtil.parseTime(start) + " – " + FluxUtil.parseTime(end))
             }
         } catch (e: JSONException) {
             e.printStackTrace()
@@ -136,16 +137,5 @@ object JsonParser {
             e.printStackTrace()
         }
         return densities
-    }
-
-    fun parseTime(timestamp: Long): String {
-        val timeZone = Calendar.getInstance().timeZone
-        var format = SimpleDateFormat("h:mma", Locale.US)
-        if (DateFormat.is24HourFormat(DensityApplication.getAppContext())) {
-            format = SimpleDateFormat("HH:mm", Locale.US)
-        }
-        format.timeZone = timeZone
-
-        return format.format(Date(timestamp * 1000)).toLowerCase(Locale.US)
     }
 }
